+++ core/client/fs/src/test/java/alluxio/client/file/MockFileSystemMasterClient.java
@@ -118,11 +118,6 @@ class MockFileSystemMasterClient implements FileSystemMasterClient {
   }
 
   @Override
-  public void updateConfig(Map<String, String> propertiesMap)
-      throws AlluxioStatusException {
-  }
-
-  @Override
   public void updateMount(AlluxioURI alluxioPath, MountPOptions options)
       throws AlluxioStatusException {
   }
