+++ core/server/master/src/main/java/alluxio/master/meta/MetaMaster.java
@@ -164,4 +164,9 @@ public interface MetaMaster extends BackupOps, Master {
    * @return the hostname of the master that did the checkpoint
    */
   String checkpoint() throws IOException;
+
+  /**
+   * @param propertiesMap properties to update
+   */
+  void updateConfig(Map<String, String> propertiesMap);
 }
